split KeyConverter and update projects
